requirements.txt
alembic==1.13.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
greenlet==3.0.2
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.2
flask-cors==4.0.0
